DOC-111 dfs-fs remove without exception, further tests to increase coverage

+++ datasafe-storage/datasafe-storage-impl-fs/src/main/java/de/adorsys/datasafe/business/impl/storage/FileSystemStorageService.java
@@ -64,6 +64,9 @@ public class FileSystemStorageService implements StorageService {
     @SneakyThrows
     @Override
     public void remove(AbsoluteLocation location) {
+        if (!objectExists(location)) {
+            return;
+        }
         Files.delete(resolve(location.location(), false));
         log.debug("deleted directory at: {}", location);
     }
